Implement open normal file
code from the click on the list view was abstracted into a local
function (instead of making variables globally). Then ) re-used the same
function to implement the regular open file command.

I am trying to avoid using OO classes for this project. So far it
is working better then expected. Proving myself that C++ is not always
about creating a class.

Well, at least not directly. Since the compiler will create a new class
for each lambda. So, we got that.